renamed the component
